Use decimals for currency arthimetic

+++ hand.history/DataObject/Action.cs
@@ -6,8 +6,8 @@ namespace hand.history.DataObject
 {
     public class Action
     {
-        public double Id { get; set; }
-        public double Amount { get; set; }
+        public decimal Id { get; set; }
+        public decimal Amount { get; set; }
 
         public Player Player { get; set; }
         public VerbType Verb { get; set; }
