+++ hand.history/DataObject/Street.cs
@@ -6,8 +6,8 @@ namespace hand.history.DataObject
 {
     public class Street
     {
-        public double Id { get; set; }
-        public double Pot { get; set; }
+        public decimal Id { get; set; }
+        public decimal Pot { get; set; }
 
         public StreetType Type { get; set; }
         public IEnumerable<Card> Community { get; set; }
