+++ hand.history/DataObject/Table.cs
@@ -9,13 +9,13 @@ namespace hand.history.DataObject
 {
     public class Table
     {
-        public double Id { get; set; }
-        public double GameId { get; set; }
-        public double TournamentId { get; set; }
-        public double BBlind { get; set; }
-        public double SBlind { get; set; }
-        public double Pot { get; set; }
-        public double Rake { get; set; }
+        public decimal Id { get; set; }
+        public decimal GameId { get; set; }
+        public decimal TournamentId { get; set; }
+        public decimal BBlind { get; set; }
+        public decimal SBlind { get; set; }
+        public decimal Pot { get; set; }
+        public decimal Rake { get; set; }
         public string Currency { get; set; }
         public string Title { get; set; }
         public string Game { get; set; }
